adds in page

+++ destroy.js
@@ -6,11 +6,27 @@ import {
   unfmt,
 } from "./deps.ts";
 
-/** destroy an item */
+/** destroy record(s) */
 export async function destroy(params) {
-  let [TableName, Key] = await Promise.all([
-    getTableName(),
-    getKey(params),
-  ]);
-  return createClient().deleteItem({ TableName, Key });
+  // destroy batch
+  if (Array.isArray(params)) {
+    let TableName = await getTableName();
+    let req = (Key) => ({ DeleteRequest: { Key } });
+    let batch = params.map(getKey).map(req);
+    let query = { RequestItems: {} };
+    query.RequestItems[TableName] = batch;
+    await createClient().batchWriteItem(query);
+    return;
+  }
+  // destroy one
+  if (params.table && params.key) {
+    let [TableName, Key] = await Promise.all([
+      getTableName(),
+      getKey(params),
+    ]);
+    await createClient().deleteItem({ TableName, Key });
+    return;
+  }
+  // destroy fail
+  throw Error("destroy_invalid");
 }
